feat: replace ExtentReports with Allure
Both reporting frameworks are designed for reporting test results. ExtentReports community version is not on Maven central, so Galenium was stuck with an old version. Allure is very much alive and a lot more flexible and extendable. 

* switch to Allure reporting
* straighten out dependency declarations
* clean up logback config
* log to files and attach to Allure
* move TestNG listeners to dedicated package

BREAKING CHANGE: Use Allure API to write to report and SLF4J loggers to log