docs: run javadoc:autofix over plugin and all modules
To have a baseline of javadoc comments with since tags Maven's Javadoc plugin was used.

Preferences was given to:
* uninformative comments over no comments. 
* live with misleading @since tags over investing manual effort 


* docs: run javadoc:fix over 1.0.0
* docs: run javadoc:fix over 2.0.0
* docs: run javadoc:fix over 3.0.0
* chore: fix pre-commit warning about unexpected key
* docs: run javadoc:fix over 4.0.0-SNAPSHOT
* revert: remove unused constants again
* docs: fix one comment, run javadoc:fix again
* docs: javadoc generation was done in specs-plugin and all modules

Creating fresh changes.xml files as changes come up. The new strategy is to have
all changes at the appropriate level in the module hierarchy and aggragate them
during site generation.