feat: make spec an interface and hide Galen better
* feat: make galen spec an interface

It makes sense to expose the interface and keep the
implementations in the module. There are two main types
of specs in Galenium: gspec file and image comparison.
These are implemented and are used now.

* refactor: separate image comparison from general spec functionality

The image comparison is very special to Galenium. It makes sense
to keep it in a separate package.

* fix: remove trivial methods from validation util

The GalenSpec interface already provides a direct way to run
the spec. There is no need to have static methods that just
call those check methods.

* refactor: reduce constructor scope

The factory methods already allow user to construct all
sensible variants. Exposing the constructor is
unnecessary noise.

* fix: rename method to better reflect function

None is too broad. Neutral implies that the correction
exists, but does not alter anything.

* test: add test for corrections

Testing the factory methods happy path.