fix: fix and clean examples and related functionality
* refactor: move example tests to subpackages according to type

To have a clearer distinction between different types of tests possible with
Galenium all examples are moved to subpackages named after the test type.

* refactor: test is sampling, so move to sampling package

* refactor: use property to switch ui-tests on and off

Less code and a speaking property name.

* feat: jetty plugin replacing AEM as SUT

Setting up a local AEM instance and deploying WCM.io Samples
is a high entry threshold, so Galenium will run its
examples against Jetty. A mirror of the sample is
served as static content.

* feat: introduce profile for using Jetty as SUT

* feat: add screenshots to test reports on all possible results

* docs: logging no longer feeds into HTML

With Allure reporting the logs are attached as text files and
there is no HTML interpretation. Logging to the command line
never benefitted from escaped HTML code.

* fix: force skipping of tests during regular build

The negation of run test property does not seem to work. Now
just hard coded configuration of failsafe plugin in profile.

* feat: let tests know whether SUT is AEM author

We need to know whether we need to login to an AEM author or
whether we can just GET our content from Jetty.

* fix: remove dead code

This is a relict from Extent report integration and no longer needed.

* feat: add scripts to start configured test run

This convention is already integrated into Galenium archetype and
makes good sense in examples.

* fix: kill AEM only test

Until there is a solution to make the ContentPathProvider
AEM agnostic, the test will just be removed.

* fix: add explicit profile for jetty run

The new default is to run the UI tests against local Jetty
and this should be reflected in the run script.

* fix: add excludes for header check and test groups

* fix: add component check to AEM group

Checking source for component is clearly an AEM specific test.

* fix: only login, if dealing with AEM author

Use the new property to decide whether to login or not.

* fix: lock down internal methods

All methods that are only called from this utility class
should be private. Also added guards to log statements.

* refactor: move image comparison to galen package

The whole image comparison is very Galen centric so it should live
in the galen package.

* fix: guard log statements

* fix: move spec related methods to dedicated type

Galen has different aspects which should be reflected in class
structure. Introducing utility class for Galen spec handling.

* fix: rename package to align with Galen

Galen uses specs as package name, so Galenium uses same
convention for Galen related functionality.

* fix: move layout to validation package

Align with Galen package naming.

* feat: add dedicated Galen reports class

Reporting and validation are separate steps in testing
and deal with different parts of the Galen API. The new
reports package and class can be home to new functionality
around Galen reports.

* fix: fix method for checking single tag

Made method functional and static. Before it was without
benefit, so it was replaced by a functioning static
version.

* feat: add example for new generated method

Now that the method is functional there should be an
example.

* style: separate methods by empty line

* feat: rename class and methods for clarity

Galen validation is always layout oriented, so removing redundant
name part from class and method names. Image comparison uses
Galen in the background, but is not the same as checking an
actual Galen spec.

* docs: remove misleading part of comment

Talking about folder names in this context does not help
understanding and can be confusing. Removing it makes the
comment better.

* style: rename listener to reflect no-op nature

The listener is better characterized by what it is doing, which in
this case is nothing or "no operation".

* refactor: extract inner classes

Since there is a validation package now, we can extract the
listeners and let them live as top level classes inside the
package.

* refactor: better method names for spec util

* fix: remove return values from setters

Setters are not a good fit for chaining. To save boiler
plate, rather use a builder or other alternative.

* fix: reduce log level because of log spam

Cache invalidation is really TRACE level. If you want the info
you can still get it, but most of the time it is noise. Also
guarding the log statement, which should have been the case
for the debug level anyway.

* perf: guard log statement

* style: remove empty line

* fix: remove return value because setter

Setters should not return chaining value.

* feat: better logging of JSoup request/response

This can be very useful, when trying to figure out, what is going
wrong during server communication.

* docs: run javadoc autofix

First executions after integration of javadoc-autofix execution.

* fix: remove duplicate classes

* fix: remove unused logger

* fix: use alternative to deprecated method

* test: add some Hamcrest based tests

This is a first stab at testing the Galenium API. Hamcrest matchers
seem like a good fit. Reusable and recombinable validation with
better error messsages.

* test: remove test

This needs a rewrite to be a nice example so just remove for
now.

* fix: fix some test examples

Test examples are not working with Jetty runs. Fix all except
image comparison. Prepare fixing image comparison.

* refactor: streamline path for checking without tags

One code path is easier to maintain. Also readability and DRY.

* style: remove dead code

* fix: fix dummy image initialization and samples

The example tests need new image samples for Jetty run and
running image comparison against a non-existent baseline was
broken.

* fix: guard debug statement

* feat: add link whenever using load functionality

Implemented a blunt way to make sure links to SUT pages are added
to report.